Add files via upload
I have done quite a lot of debugging, since this commit is for two days of work. First thing is that you should be able to use any csv file no matter the size or amounts of students ,teachers and served. I have also added an option in the settings to change what type of csv file you are using, with debugging to make sure it's a csv file. Lastly I have added more buttons into my gui to make it easier to use, if you are not using the menu items for moving from page to page.